Changed to explicit block ID calls
find_blocks("") no longer returns a list of all blocks in the
design, so the fix here is temporary as the final "else" will
never evaluate to find other blocks in the system.  This is
a new bug until we can find a way to read all detected blocks
from uhd::device3.

Signed-off-by: Thomas Goodwin <btgoodwin@geontech.com>